% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbor.R
\name{get_neighbor}
\alias{get_neighbor}
\title{Get neighbor in the geometry}
\usage{
get_neighbor(shapefile, tipo = "Queen")
}
\arguments{
\item{shapefile}{A spatial object of class \code{sfc_MULTIPOLYGON}}

\item{tipo}{type of neighbor.}
}
\value{
An object \code{vizi}
}
\description{
\code{get_neighbor()} get neighbors from an spatial object with class \code{sfc_MULTIPOLYGON} and returns a list object with class equal to "vizi".
}
\examples{
\dontrun{
# Import
nc <- sf::read_sf(system.file("gpkg/nc.gpkg", package = "sf"))

# Get the neighbor
l_nb <- get_neighbor(nc, "Queen")

# Visualize
view_connections(l_nb)

view_neighbor(l_nb)

}
}
