% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_neighbor.R
\name{view_neighbor}
\alias{view_neighbor}
\title{View the neighborhood}
\usage{
view_neighbor(nbr)
}
\arguments{
\item{nbr}{A neighbor object create by \code{get_neighbor()} function.}
}
\value{
An object \code{ggplot}
}
\description{
\code{{view_neighbor()}} get neighbors from an neighbor object and returns a ggplot object.
}
\examples{
\dontrun{
# Import
nc <- sf::read_sf(system.file("gpkg/nc.gpkg", package = "sf"))

# Get the neighbor
l_nb <- get_neighbor(nc, "Queen")

# Visualize
view_connections(l_nb)

view_neighbor(l_nb)

}
}
